--- client/src/component/molecule/ButtonDiv.tsx
@@ -1,28 +1,0 @@
-import React from 'react'
-import A from '@atom'
-import { ButtonType } from '@atom/Button/index'
-import { TextType } from '@atom/Text/index'
-
-namespace ButtonDivType {
-  export interface Props {
-    buttonStyle?: ButtonType.StyleAttributes
-    textStyle?: TextType.StyleAttributes
-    children?: React.ReactChild
-    onClick?: () => void
-  }
-}
-
-const ButtonDiv = ({
-  buttonStyle,
-  textStyle,
-  children,
-  onClick,
-}: ButtonDivType.Props) => {
-  return (
-    <A.Button customStyle={buttonStyle} onClick={onClick}>
-      <A.Text customStyle={textStyle}>{children}</A.Text>
-    </A.Button>
-  )
-}
-
-export default ButtonDiv
