+++ client/src/component/molecule/ButtonDiv/index.ts
@@ -1,0 +1,12 @@
+import { ReactChild } from 'react'
+import { ButtonType } from '@atom/Button'
+import { TextType } from '@atom/Text'
+
+export { default } from './ButtonDiv'
+
+export interface ButtonDivProps {
+  buttonStyle?: ButtonType.StyleAttributes
+  textStyle?: TextType.StyleAttributes
+  children?: ReactChild
+  onClick?: () => void
+}
